Add an extended STS over a different base monad
Add an example of extended STS
